refactor(app): deduplicate eligible students sheet
